backend/requirements.txt
# Web フレームワーク
Django==5.2
djangorestframework==3.14.0

# PostgreSQL ドライバ（開発では psycopg2-binary の方が安全）
psycopg2-binary>=2.9

# WSGI サーバー
uWSGI>=2.0.20
